scripts/requirements.txt
paho-mqtt>=2.0.0
websocket-client>=1.6.0
